Move Arguments to own folder
